+++ src/frontend/src/common/graphql/queries.ts
