+++ src/frontend/src/common/graphql/types.ts
