+++ src/frontend/src/common/platform.ts
