+++ src/frontend/src/index.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from "react-dom/client";
 // import React from 'react';
 import App from "./App";
-import { FlashbarProvider } from "./components/notifications";
+import { FlashbarProvider } from "./common/components/notifications";
 
 const root = ReactDOM.createRoot(document.getElementById("root")!);
 root.render(
